# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koperasi App API",
        "description": "RESTful API for Koperasi App mobile applications and integrations",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pinjaman": {
            "get": {
                "tags": [
                    "Pinjaman"
                ],
                "summary": "List user loans",
                "description": "Display a listing of the resource.",
                "operationId": "34915626407efaf8d9c7511cad357794",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of loans",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Pinjaman"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Anggota not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pinjaman"
                ],
                "summary": "Submit a new loan application",
                "description": "Store a newly created resource in storage.",
                "operationId": "22b9e5d59082a02fef5c88c9e3b36644",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nominal",
                                    "tenor_bulan",
                                    "keperluan"
                                ],
                                "properties": {
                                    "nominal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000000,
                                        "minimum": 1000000
                                    },
                                    "tenor_bulan": {
                                        "type": "integer",
                                        "example": 12,
                                        "maximum": 120,
                                        "minimum": 1
                                    },
                                    "keperluan": {
                                        "type": "string",
                                        "example": "Renovasi rumah",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Loan application created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pinjaman"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Only active members can submit applications"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pinjaman/{id}": {
            "get": {
                "tags": [
                    "Pinjaman"
                ],
                "summary": "Get loan details",
                "description": "Display the specified resource.",
                "operationId": "44e88c06f2fd2d1a4d0f72cc06687255",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Loan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Loan details with installments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pinjaman"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Unauthorized - loan does not belong to you"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pinjaman/{id}/cek-nominal": {
            "post": {
                "tags": [
                    "Pinjaman"
                ],
                "summary": "Check loan eligibility for a nominal amount",
                "description": "Check loan eligibility/nominal",
                "operationId": "6c5d52bca564ea8254592ec6014b0668",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Loan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nominal"
                                ],
                                "properties": {
                                    "nominal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000000,
                                        "minimum": 1000000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Eligibility check result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "eligible": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "max_nominal": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 960000
                                        },
                                        "current_nominal": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 5000000
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Loan amount exceeds maximum eligible amount"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pinjaman/{id}/simulasi": {
            "post": {
                "tags": [
                    "Pinjaman"
                ],
                "summary": "Simulate loan repayment schedule",
                "description": "Simulate loan",
                "operationId": "273499ea4897298091189fdc27055358",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Loan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nominal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000000,
                                        "minimum": 1000000
                                    },
                                    "tenor_bulan": {
                                        "type": "integer",
                                        "example": 12,
                                        "maximum": 120,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Loan simulation result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nominal": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "tenor_bulan": {
                                            "type": "integer"
                                        },
                                        "pokok_per_bulan": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "bunga_per_bulan": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "angsuran_per_bulan": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "total_bunga": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "total_pembayaran": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "rincian_angsuran": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "cicilan_ke": {
                                                        "type": "integer"
                                                    },
                                                    "pokok": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "bunga": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "total": {
                                                        "type": "number",
                                                        "format": "float"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Pinjaman": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "anggota_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pengaju_user_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "nominal": {
                        "type": "string",
                        "example": "5000000.00"
                    },
                    "tenor_bulan": {
                        "type": "integer",
                        "example": 12
                    },
                    "keperluan": {
                        "type": "string",
                        "example": "Renovasi rumah"
                    },
                    "snapshot_bank": {
                        "type": "string",
                        "example": "BCA",
                        "nullable": true
                    },
                    "snapshot_no_rekening": {
                        "type": "string",
                        "example": "1234567890",
                        "nullable": true
                    },
                    "snapshot_atas_nama": {
                        "type": "string",
                        "example": "John Doe",
                        "nullable": true
                    },
                    "persentase_bunga": {
                        "type": "string",
                        "example": "1.00"
                    },
                    "status": {
                        "type": "string",
                        "example": "diajukan",
                        "enum": [
                            "diajukan",
                            "approved_bendahara",
                            "approved_ketua",
                            "aktif",
                            "ditolak",
                            "lunas"
                        ]
                    },
                    "cair_oleh_bendahara": {
                        "type": "boolean",
                        "example": false
                    },
                    "sudah_pakai_privilege_reloan": {
                        "type": "boolean",
                        "example": false
                    },
                    "sudah_pakai_percepatan": {
                        "type": "boolean",
                        "example": false
                    },
                    "tanggal_pengajuan": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "tanggal_pencairan": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "disetujui_pada": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "versi_syarat": {
                        "type": "string",
                        "nullable": true
                    },
                    "ip_address_setuju": {
                        "type": "string",
                        "nullable": true
                    },
                    "user_agent_setuju": {
                        "type": "string",
                        "nullable": true
                    },
                    "catatan_bendahara": {
                        "type": "string",
                        "nullable": true
                    },
                    "catatan_ketua": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "anggota": {
                        "$ref": "#/components/schemas/Anggota"
                    },
                    "pengaju": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}